requirements.txt
Flask
Flask-SocketIO
requests
numpy
scipy
python-dotenv  # API anahtarını .env dosyasında saklamak için